--- client/src/components/RelatedProducts/RelatedProducts.css
@@ -1,41 +1,0 @@
-.related-container {
-  margin: 20px 0;
-  color: rgb(255, 255, 255);
-  background-color: rgba(182, 173, 173, 0.9);
-  border-radius: 10px;
-}
-
-.related-container h3 {
-  font-size: 22px;
-  color: black;
-  background-color: rgb(66, 64, 64);
-  border-top-left-radius: 10px;
-  border-top-right-radius: 10px;
-  color: rgb(234, 234, 234);
-  padding: 5px;
-  text-align: left;
-  padding-left: 10px;
-  font-weight: 500;
-}
-
-.related-container h4 {
-  font-size: 20px;
-  text-align: left;
-  margin-left: 10px;
-  font-weight: 400;
-  overflow-y: hidden;
-  margin: 1px;
-}
-
-.related-container {
-  background-color: grey;
-  position: absolute;
-  left: 0;
-}
-
-@media only screen and (max-width: 1200px) {
-  .related-container {
-    position: relative;
-    margin: 20 auto;
-  }
-}
